#module2-task1: added propTypes
